requirements.txt
appnope==0.1.0
backcall==0.1.0
boto3==1.7.19
botocore==1.10.19
certifi==2018.4.16
chardet==3.0.4
cryptocompy==0.1.1.dev1
cycler==0.10.0
decorator==4.3.0
docutils==0.14
dropbox==8.8.1
ffnet==0.8.3
h5py==2.7.1
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jmespath==0.9.3
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
nltk==3.3
numpy==1.14.3
oauthlib==2.0.7
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
scipy==1.1.0
searchtweets==1.7.0
simplegeneric==0.8.1
six==1.11.0
textblob==0.15.1
traitlets==4.3.2
tweet-parser==1.13.0
urllib3==1.22
wcwidth==0.1.7
